+++ lib/rbcli/features/autoupdate/gem_updater.rb
@@ -18,56 +18,45 @@
 #     For questions regarding licensing, please contact andrew@blacknex.us       #
 ##################################################################################
 
-require 'octokit'
+require 'rbcli/features/autoupdate/common/autoupdate'
+require 'net/http'
+require 'json'
 
 module Rbcli::Autoupdate
-	class GithubUpdater
+	class GemUpdater
 		include Common
 
-		def initialize reponame, access_token, enterprise_hostname, force_update, message
-			@reponame = reponame
+		def initialize gemname, force_update, message
+			@gemname = gemname
+			@uri = URI.parse "https://rubygems.org/api/v1/versions/#{gemname}/latest.json"
 			@force_update = force_update
 			@message = message
-
-			access_token = Rbcli.config.key?(:github_updatecheck) ? (Rbcli.config[:github_updatecheck][:access_token] || access_token) : access_token
-			enterprise_hostname = Rbcli.config.key?(:github_updatecheck) ? (Rbcli.config[:github_updatecheck][:enterprise_hostname] || enterprise_hostname) : enterprise_hostname
-
-			# Enterprise Connection
-			if enterprise_hostname
-				Octokit.configure do |c|
-					c.api_endpoint = "https://#{enterprise_hostname}/api/v3/"
-				end
-				# Public Github Connection
-			end
-			#OAuth connection - not used
-			#@client = Octokit::Client.new :client_id => client_id, :client_secret => client_secret
-			@client = Octokit::Client.new(:access_token => access_token)
 		end
 
 		def get_latest_version
 			begin
-				@client.repo(@reponame).rels[:tags].get.data.map{ |t| t[:name] }[0].sub(/^[v]*/,"")
-			rescue Faraday::ConnectionFailed => e
-				# This is to capture connection errors without bothering the user.
+				response = Net::HTTP.get(@uri)
+				JSON.parse(response)['version']
+			rescue SocketError => e
+				# Capture connection errors
 			end
-
 		end
 
 		def update_message
-			"Please check the github repo #{@reponame} for instructions."
+			"Please run `gem update #{@gemname}` to upgrade to the latest version. You can see it at: https://rubygems.org/gems/#{@gemname}/versions/#{@latest_version}"
 		end
 
-		def self.save_defaults
-			Rbcli::Config::add_categorized_defaults :github_update, 'Automatically Check for Updates from GitHub', {
-					access_token: {
-							description: 'Access token for GitHub API. This is only required for private repos. For help, see: https://help.github.com/articles/creating-a-personal-access-token-for-the-command-line/',
-							value: nil
-					},
-					enterprise_hostname: {
-							description: 'Hostname for GitHub Enterprise. Leave as null (~) for public GitHub.',
-							value: nil
-					}
-			}
-		end
+		# def self.save_defaults
+		# 	Rbcli::Config::add_categorized_defaults :gem_update, 'Automatically Check for Updates from RubyGems', {
+		# 			access_token: {
+		# 					description: 'Access token for GitHub API. This is only required for private repos. For help, see: https://help.github.com/articles/creating-a-personal-access-token-for-the-command-line/',
+		# 					value: nil
+		# 			},
+		# 			enterprise_hostname: {
+		# 					description: 'Hostname for GitHub Enterprise. Leave as null (~) for public GitHub.',
+		# 					value: nil
+		# 			}
+		# 	}
+		# end
 	end
 end