+++ docs-src/docs/imported/changelog.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## 0.2.2 (Aug 22, 2018)
+
+### Features
+
+### Bugfixes
+
+* Fixed a bug that caused the logger's target and level not to be configured properly via the Configurate block.
+
+### Improvements
+
+* Lazy-loading has been implemented in optional modules such as autoupdates, remote storage, etc. This means that if you do not enable them in the code, they will not be loaded into memory. This significantly improves loding times for applications.
+* Abstraction system created for configuration. This has significantly simplified the existing codebase and makes future development easier.
+* Deprecation warning system added. This allows for RBCli contributors to notify users of breaking changes that may impact their code.
+* Folder structure has been simplified to ease development.
+* Much of the code has been refactored.
+
+### Deprecations/Changes
+
+* The `Rbcli` module is now `RBCli` to better match the branding. The original `Rbcli` module will still work for this current release, with a warning, but future releases will require code changes. 
+* Hooks are now defined under the `RBCli.Configurate.hooks` block instead of `RBCli.Configurate.me`.
+* The logger is now silent by default. To enable it, it must be configured either via the `Configurate` block or via the user's config file.
+
+
 ## 0.2.1 (Aug 8, 2018)
 
 ### Features